arbitrage_platform/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
cryptography
requests
gunicorn
dj_database_url
celery
redis
django-redis
python-json-logger
